+++ tests/tests/models/test_dummy.py
@@ -1,14 +1,15 @@
+import math
 import os
 import posixpath
-import math
 from contextlib import contextmanager
+from pathlib import Path
 
 import pytest
 from django.core.files import File
 from django.db.models.fields import Field
 
-from app.models.dummy import *
 from app.models.base import FileExample
+from app.models.dummy import *
 from paper_uploads import signals
 from paper_uploads.files import VariationFile
 from paper_uploads.variations import PaperVariation
@@ -688,6 +689,24 @@ class TestFileFieldResourceAttach:
         finally:
             resource.delete_file()
 
+    def test_file_as_string(self):
+        with self.get_resource() as resource:
+            resource.attach(NASA_FILEPATH)
+
+            assert resource.basename == "milky-way-nasa"
+            assert resource.extension == "jpg"
+            assert resource.size == self.resource_size
+            assert resource.checksum == self.resource_checksum
+
+    def test_file_as_path(self):
+        with self.get_resource() as resource:
+            resource.attach(Path(NASA_FILEPATH))
+
+            assert resource.basename == "milky-way-nasa"
+            assert resource.extension == "jpg"
+            assert resource.size == self.resource_size
+            assert resource.checksum == self.resource_checksum
+
     def test_file(self):
         with self.get_resource() as resource:
             with open(NASA_FILEPATH, 'rb') as fp:
@@ -700,31 +719,63 @@ class TestFileFieldResourceAttach:
 
     def test_django_file(self):
         with self.get_resource() as resource:
-            with open(NASA_FILEPATH, 'rb') as fp:
-                file = File(fp, name='milky-way-nasa.jpg')
+            with open(NASA_FILEPATH, "rb") as fp:
+                file = File(fp, name="milky-way-nasa.jpg")
                 resource.attach(file)
 
-            assert resource.basename == 'milky-way-nasa'
-            assert resource.extension == 'jpg'
+            assert resource.basename == "milky-way-nasa"
+            assert resource.extension == "jpg"
+            assert resource.size == self.resource_size
+            assert resource.checksum == self.resource_checksum
+
+    def test_django_file_with_relative_path(self):
+        with self.get_resource() as resource:
+            with open(NASA_FILEPATH, "rb") as fp:
+                file = File(fp, name="photos/milky-way-nasa.jpg")
+                resource.attach(file)
+
+            assert "/photos/" not in resource.name
+            assert resource.basename == "milky-way-nasa"
+            assert resource.extension == "jpg"
             assert resource.size == self.resource_size
             assert resource.checksum == self.resource_checksum
 
     def test_override_name(self):
         with self.get_resource() as resource:
-            with open(NASA_FILEPATH, 'rb') as fp:
-                resource.attach(fp, name='overwritten.jpg')
+            with open(NASA_FILEPATH, "rb") as fp:
+                resource.attach(fp, name="overwritten.jpg")
 
-            assert resource.basename == 'overwritten'
-            assert resource.extension == 'jpg'
+            assert "/photos/" not in resource.name
+            assert resource.basename == "overwritten"
+            assert resource.extension == "jpg"
+
+    def test_override_name_with_relative_path(self):
+        with self.get_resource() as resource:
+            with open(NASA_FILEPATH, "rb") as fp:
+                resource.attach(fp, name="photos/overwritten.jpg")
+
+            assert "/photos/" not in resource.name
+            assert resource.basename == "overwritten"
+            assert resource.extension == "jpg"
 
     def test_override_django_name(self):
         with self.get_resource() as resource:
-            with open(NASA_FILEPATH, 'rb') as fp:
-                file = File(fp, name='not_used.png')
-                resource.attach(file, name='overwritten.jpg')
+            with open(NASA_FILEPATH, "rb") as fp:
+                file = File(fp, name="not_used.png")
+                resource.attach(file, name="overwritten.jpg")
 
-            assert resource.basename == 'overwritten'
-            assert resource.extension == 'jpg'
+            assert resource.basename == "overwritten"
+            assert resource.extension == "jpg"
+
+    def test_override_django_name_with_relative_path(self):
+        with self.get_resource() as resource:
+            with open(NASA_FILEPATH, "rb") as fp:
+                file = File(fp, name="photos/not_used.png")
+                resource.attach(file, name="overwritten.jpg")
+
+            assert "/photos/" not in resource.name
+            assert resource.basename == "overwritten"
+            assert resource.extension == "jpg"
 
     def test_wrong_extension(self):
         with self.get_resource() as resource: