+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - **Note**: `paper_uploads.cloudinary` will be moved to a separate package.
 ### Features
 - `SizeValidator` is now deprecated in favor of `MaxSizeValidator`.
+- Allow `str` and `Path` as the parameter of `attach()` method.
 
 ## [0.8.0rc2](https://github.com/dldevinc/paper-uploads/tree/v0.8.0rc2) - 2021-02-01
 ### Features